% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.r
\docType{data}
\name{bibliography}
\alias{bibliography}
\title{bib file consisting of a large character vector}
\format{An indexed character vector with 216 lines}
\usage{
data(bibliography)
}
\description{
This dataset represents how a raw bib file will be processed
when it is read into R. By their nature, .bib files must first
be read in and then worked upon.
}
\keyword{datasets,bibliography}
