% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elim.r
\name{elim}
\alias{elim}
\title{Remove front parts of copied paths}
\usage{
elim(a)
}
\arguments{
\item{a}{path to files}
}
\value{
a cleansed path to a folder of documents
}
\description{
Remove front parts of copied paths
}
