% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstocr.r
\name{jstocr}
\alias{jstocr}
\title{merge metadata and ocr text informatics from JSTOr's D4F (Documents for Research). The function currently returns a list of information for downstream further processing.}
\usage{
jstocr(a1 = path, metadata = T, ngram1 = F, ngram2 = F,
  ngram3 = F, ocr = F)
}
\arguments{
\item{a1}{The path to the unzipped folder holding all of the subfolders provided by JSTOR D4R}

\item{metadata}{A binary variable (T/F) indicating whether the main folder has a metadata subfolder. If there is no folder detected, it will produce an error message. Defaults to T.}

\item{ngram1}{A binary variable (T/F) indicating the presence of an ngram1 subfolder. If false, the function returns an empty vector.}

\item{ngram2}{A binary variable (T/F) indicating the presence of an ngram2 subfolder. If false, the function returns an empty vector.}

\item{ngram3}{A binary variable (T/F) indicating the presence of an ngram3 subfolder. If false, the function returns an empty vector.}

\item{ocr}{A binary varialbe (T/F) indicating the presence of an ocr subfolder. If false, the function returns an empty vector.}
}
\value{
a list of information
}
\description{
merge metadata and ocr text informatics from JSTOr's D4F (Documents for Research). The function currently returns a list of information for downstream further processing.
}
\examples{
jstocr(a1 = path, ocr = T)
}
