% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstor_ocr.r
\name{jstor_ocr}
\alias{jstor_ocr}
\title{Take metadata and ocr text from D4R and merge it into xml files.}
\usage{
jstor_ocr(a1 = input_path, b1 = output_path)
}
\arguments{
\item{a1}{The path to the unzipped Jstor folder.}

\item{b1}{The path to a folder that will contain the output files.}
}
\value{
xml files with merged metadata and machine readable text.
}
\description{
Take metadata and ocr text from D4R and merge it into xml files.
}
\examples{
jstor_ocr(a = "path/to/d4r/files", b = "destination/folder/path")
}
