% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuter_spanish.r
\name{neuter_spanish}
\alias{neuter_spanish}
\title{Take metadata and ocr text from D4R and merge it into xml files.}
\usage{
neuter_spanish(x)
}
\arguments{
\item{x}{A character vector with Spanish in it. At the moment has not been tested on ascii files or latin encoding.}
}
\value{
Neutered nouns and adjectives in Spanish. This package function was developed for a particular purpose, and

the project goal should be closely considered before implementing. The function is currently in beta testing.
}
\description{
Take metadata and ocr text from D4R and merge it into xml files.
}
