% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangulate.R
\name{rectangulate}
\alias{rectangulate}
\title{A Matrix Literature Review Data Frame}
\usage{
rectangulate(a = path_to_bib)
}
\arguments{
\item{a}{A .bib file}
}
\value{
A dataframe to export as a foundation for the matrix literature review
}
\description{
A Matrix Literature Review Data Frame
}
