% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstor_ocr.r
\name{jstor_ocr}
\alias{jstor_ocr}
\title{Take metadata and ocr text from D4R and merge it into xml files.}
\usage{
jstor_ocr(a1 = input_path, b1 = output_path)
}
\arguments{
\item{a1}{The path to the unzipped Jstor folder.}

\item{b1}{The path to a folder that will contain the output files.}
}
\value{
Produces .xml files with concatenated metadata and machine readable text, and a quality control message indicating parsimony between indexed file names. Assumes a1 = unzipped file location. Note well: Due to inconsistent text tagging from file sources originating from JSTOR, may throw error messages when processing.
}
\description{
Take metadata and ocr text from D4R and merge it into xml files.
}
\examples{
jstor_ocr(a1 = "path/to/d4r/files", b1 = "destination/folder/path")
}
